backend/requirements.txt
Flask==2.3.2
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.4.0
langgraph==0.0.26
pymongo==4.4.1
langchain==0.1.0
